updates to matches and wakeup
This makes add_match take a "rule" simiar to filters to set matces.
ebus_proxy now sets a match on add_sinal_handler

This also sets a one second wakeup_timer to kick the main dispatch in
dbus. I never thought this would be needed but so far it appears to
fix some of the issues